+++ humanMode.api.js
@@ -1,0 +1,13 @@
+// --- humanMode.api.js ---
+const VERSION = 'v1.0.0';
+/*
+ * Human Mode Facade
+ * - createPlayerContainers({ dom, config })
+ * - initPlayersSequentially({ Scheduler, Utils, PlayerController, Config, Guards, seqInit, Logger, players })
+ */
+
+export function getVersion() { return VERSION; }
+export function createPlayerContainers({ dom, config }) {}
+export async function initPlayersSequentially(args) {}
+
+// --- End Of File ---